gopython3/requirements.txt
django-configurations==0.6
dj-database-url==0.2.2
South==0.8.2

hammock==0.2.4
requests==2.0.0
httpretty==0.5.9
distlib==0.1.2
-e git://github.com/jezdez/django_compressor.git@c961dadf6eaba93a9cdadc569c697828e9afe345#egg=django-compressor

django-model-utils==1.5.0
jsonfield==0.9.19
djangorestframework==2.3.8
Markdown==2.3.1
django-celery==3.0.23
celery==3.0.23
pytz==2013d
billiard==2.7.3.32
python-dateutil==2.1
kombu==2.5.14
anyjson==0.3.3
six==1.4.1

# Django 1.6b4
# Should be last, just in case
https://www.djangoproject.com/download/1.6b4/tarball/
